requirements/test.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements/test.txt requirements/test.in
#
arrow==1.2.2
    # via jinja2-time
attrs==21.4.0
    # via
    #   flake8-bugbear
    #   flake8-mypy
binaryornot==0.4.4
    # via cookiecutter
black==21.12b0
    # via -r requirements/test.in
certifi==2021.10.8
    # via requests
chardet==4.0.0
    # via binaryornot
charset-normalizer==2.0.12
    # via requests
click==8.0.4
    # via
    #   -r requirements/common.in
    #   black
    #   cookiecutter
cookiecutter==1.7.3
    # via -r requirements/common.in
coverage[toml]==5.5
    # via -r requirements/test.in
decorator==5.1.1
    # via validators
flake8==3.9.2
    # via
    #   -r requirements/test.in
    #   flake8-bugbear
    #   flake8-docstrings
    #   flake8-isort
    #   flake8-mypy
flake8-bugbear==21.4.3
    # via -r requirements/test.in
flake8-docstrings==1.6.0
    # via -r requirements/test.in
flake8-isort==4.0.0
    # via -r requirements/test.in
flake8-mypy==17.8.0
    # via -r requirements/test.in
idna==3.3
    # via requests
isort==5.9.3
    # via
    #   -r requirements/test.in
    #   flake8-isort
jinja2==3.0.3
    # via
    #   cookiecutter
    #   jinja2-time
jinja2-time==0.2.0
    # via cookiecutter
markupsafe==2.1.0
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy==0.931
    # via
    #   -r requirements/test.in
    #   flake8-mypy
mypy-extensions==0.4.3
    # via
    #   black
    #   mypy
pathspec==0.9.0
    # via black
platformdirs==2.5.1
    # via black
poyo==0.5.0
    # via cookiecutter
pycodestyle==2.7.0
    # via flake8
pydocstyle==6.1.1
    # via flake8-docstrings
pyflakes==2.3.1
    # via flake8
python-dateutil==2.8.2
    # via arrow
python-slugify==6.1.1
    # via cookiecutter
requests==2.27.1
    # via cookiecutter
six==1.16.0
    # via
    #   cookiecutter
    #   python-dateutil
    #   validators
snowballstemmer==2.2.0
    # via pydocstyle
testfixtures==6.18.5
    # via flake8-isort
text-unidecode==1.3
    # via python-slugify
toml==0.10.2
    # via coverage
tomli==1.2.3
    # via
    #   black
    #   mypy
types-python-slugify==0.1.2
    # via -r requirements/common.in
typing-extensions==4.1.1
    # via
    #   black
    #   mypy
urllib3==1.26.8
    # via requests
validators==0.18.2
    # via -r requirements/common.in
